Add SCRYPTED_HOSTNAME setting
When using Scrypted with https and a hostname, video clips fail to
display because the URL that Scrypted constructs uses an IP address,
which is typically not included in the TLS certificate CN/Subject
Alternative Name.

Add support for SCRYPTED_HOSTNAME setting which, if set, is used to
construct video clip URLs instead of the server's IP address.

Fixes #367, #445.